feat: Deprecate defaultSinks and appendDefaultInterceptors
