+++ packages/worker/src/index.ts
@@ -17,8 +17,22 @@ export {
   PrometheusMetricsExporter,
   TelemetryOptions,
 } from '@temporalio/core-bridge';
-export { LoggerSinks } from '@temporalio/workflow';
-export { ActivityInboundLogInterceptor, activityLogAttributes } from './activity-log-interceptor';
+export {
+  /**
+   * @deprecated Do not use LoggerSinks directly. To log from Workflow code, use the `log` object
+   *             exported by the `@temporalio/workflow` package. To capture log messages emitted
+   *             by Workflow code, set the {@see Runtime.logger} property.
+   */
+  LoggerSinks,
+} from '@temporalio/workflow';
+export {
+  /**
+   * @deprecated Extending `ActivityInboundLogInterceptor` is deprecated. To customize activity log attributes, simply
+   *             register a custom `WorkflowInterceptors` that intercepts the `outbound.getLogAttributes()` method.
+   */
+  ActivityInboundLogInterceptor,
+  activityLogAttributes,
+} from './activity-log-interceptor';
 export { NativeConnection as NativeConnection } from './connection';
 export { NativeConnectionOptions, RequiredNativeConnectionOptions, TLSConfig } from './connection-options';
 export { startDebugReplayer } from './debug-replayer';
@@ -39,9 +53,16 @@ export {
   WorkerStatus,
 } from './worker';
 export {
-  appendDefaultInterceptors,
+  /**
+   * @deprecated Including `appendDefaultInterceptors()` in the worker options is no longer required.
+   */
+  appendDefaultInterceptors, // eslint-disable-line deprecation/deprecation
   CompiledWorkerOptions,
-  defaultSinks,
+  /**
+   * @deprecated Including `defaultSinks()` in the worker options is no longer required. To configure
+   *             a custom logger, set the {@see Runtime.logger} property instead.
+   */
+  defaultSinks, // eslint-disable-line deprecation/deprecation
   ReplayWorkerOptions,
   WorkerOptions,
   WorkflowBundle,
@@ -51,17 +72,23 @@ export {
 } from './worker-options';
 export { ReplayError, ReplayHistoriesIterable, ReplayResult } from './replay';
 export {
+  /**
+   * @deprecated Extending `WorkflowInboundLogInterceptor` is deprecated. To customize workflow log attributes, simply
+   *             register a custom `WorkflowInterceptors` that intercepts the `outbound.getLogAttributes()` method.
+   */
   WorkflowInboundLogInterceptor, // eslint-disable-line deprecation/deprecation
+  /**
+   * @deprecated Extending `WorkflowLogInterceptor` is deprecated. To customize workflow log attributes, simply
+   *             register a custom `WorkflowInterceptors` that intercepts the `outbound.getLogAttributes()` method.
+   */
   WorkflowLogInterceptor,
+  /**
+   * @deprecated This function is meant for internal usage. Don't use it.
+   */
   workflowLogAttributes,
 } from './workflow-log-interceptor';
-export {
-  BundleOptions,
-  bundleWorkflowCode,
-  defaultWorkflowInterceptorModules,
-  /**
-   * @deprecated Use `defaultWorkflowInterceptorModules` instead
-   */
-  defaultWorkflowInterceptorModules as defaultWorflowInterceptorModules,
-  WorkflowBundleWithSourceMap,
-} from './workflow/bundler';
+export { BundleOptions, bundleWorkflowCode, WorkflowBundleWithSourceMap } from './workflow/bundler';
+/**
+ * @deprecated Including `defaultWorkflowInterceptorModules` in BundlerOptions.workflowInterceptorModules is no longer required.
+ */
+export const defaultWorkflowInterceptorModules = [];
