+++ packages/worker/src/workflow/threaded-vm.ts
@@ -11,7 +11,7 @@
 
 import { Worker as NodeWorker } from 'node:worker_threads';
 import { coresdk } from '@temporalio/proto';
-import { IllegalStateError, SinkCall } from '@temporalio/workflow';
+import { IllegalStateError, type SinkCall } from '@temporalio/workflow';
 import { UnexpectedError } from '@temporalio/core-bridge';
 import {
   WorkflowBundleWithSourceMapAndFilename,
