+++ packages/worker/src/workflow/vm-shared.ts
@@ -4,7 +4,7 @@ import { SourceMapConsumer } from 'source-map';
 import { cutoffStackTrace, IllegalStateError } from '@temporalio/common';
 import { coresdk } from '@temporalio/proto';
 import type { WorkflowInfo, FileLocation } from '@temporalio/workflow';
-import { SinkCall } from '@temporalio/workflow/lib/sinks';
+import { type SinkCall } from '@temporalio/workflow/lib/sinks';
 import * as internals from '@temporalio/workflow/lib/worker-interface';
 import { Activator } from '@temporalio/workflow/lib/internals';
 import { partition } from '../utils';
